refactor: move view and store components into appropiate folders
