+++ src/stores/StakePanelStore.tsx
